Improve code quality and remove unnecessary files
