+++ src/main/java/Duke/Duke.java
@@ -1,35 +1,137 @@
 package duke;
 
+import javafx.application.Application;
+import javafx.application.Platform;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.Region;
+import javafx.scene.layout.VBox;
+import javafx.stage.Stage;
+import javafx.scene.image.Image;
 import task.Task;
-
-import java.io.IOException;
 
 import java.util.ArrayList;
 
-public class Duke {
+public class Duke extends Application {
+
+    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
+    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
+
+    private ScrollPane scrollPane;
+    private VBox dialogContainer;
+    private TextField userInput;
+    private Button sendButton;
+    private Scene scene;
+
+    private static Storage storage = new Storage();
+    private static ArrayList<Task> list = storage.load();
+    private static CommandList commandList = new CommandList();
 
     /**
-     * Main method of Duke program.
+     * Main entry point for the GUI.
      *
-     * @param args command line arguments
+     * @param stage Stage object for the scene to populate
      */
-    public static void main(String[] args) {
+    @Override
+    public void start(Stage stage) {
+        //Step 1. Setting up required components
 
-        Ui ui = new Ui();
-        ui.displayGreeting();
+        //The container for the content of the chat to scroll.
+        scrollPane = new ScrollPane();
+        dialogContainer = new VBox();
+        scrollPane.setContent(dialogContainer);
 
-        Storage storage = new Storage();
-        ArrayList<Task> list = storage.load();
+        userInput = new TextField();
+        sendButton = new Button("Send");
 
-        Parser parser = new Parser();
-        CommandList commandList = new CommandList();
+        AnchorPane mainLayout = new AnchorPane();
+        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
 
-        while (!parser.isTerminated) {
-            try {
-                commandList.execute(parser.readInput(), parser, list, storage);
-            } catch (IOException e) {
-                System.out.println("Unable to read input!");
-            }
+        scene = new Scene(mainLayout);
+
+        stage.setScene(scene);
+        stage.show();
+
+        //Step 2. Formatting the window to look as expected
+        stage.setTitle("Duke");
+        stage.setResizable(false);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(400.0);
+
+        mainLayout.setPrefSize(400.0, 600.0);
+
+        scrollPane.setPrefSize(385, 535);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+
+        scrollPane.setVvalue(1.0);
+        scrollPane.setFitToWidth(true);
+
+        // You will need to import `javafx.scene.layout.Region` for this.
+        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
+
+        userInput.setPrefWidth(325.0);
+
+        sendButton.setPrefWidth(55.0);
+
+        AnchorPane.setTopAnchor(scrollPane, 1.0);
+
+        AnchorPane.setBottomAnchor(sendButton, 1.0);
+        AnchorPane.setRightAnchor(sendButton, 1.0);
+
+        AnchorPane.setLeftAnchor(userInput , 1.0);
+        AnchorPane.setBottomAnchor(userInput, 1.0);
+
+        //Part 3. Add functionality to handle user input.
+        sendButton.setOnMouseClicked((event) -> {
+            handleUserInput();
+        });
+
+        userInput.setOnAction((event) -> {
+            handleUserInput();
+        });
+
+        //Scroll down to the end every time dialogContainer's height changes.
+        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
+    }
+
+    /**
+     * Method to be called when application is called to stop.
+     */
+    @Override
+    public void stop() {
+        System.out.println("STOP");
+        Platform.exit();
+        System.exit(0);
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    private void handleUserInput() {
+        String input = userInput.getText();
+
+        if (input.equals("bye")) {
+            Platform.exit();
+            System.exit(0);
         }
+
+        String response = getResponse(input);
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(input, user),
+                DialogBox.getDukeDialog(response, duke)
+        );
+        userInput.clear();
+    }
+
+    /**
+     * Gets response from Duke in response to user's input.
+     */
+    protected String getResponse(String input) {
+        return commandList.execute(input, list, storage);
     }
 }
