+++ src/main/java/Duke/Main.java
@@ -13,7 +13,7 @@ import javafx.stage.Stage;
  */
 public class Main extends Application {
 
-    private TestUi testUi = new TestUi();
+    private Duke duke = new Duke();
 
     @Override
     public void start(Stage stage) {
@@ -22,7 +22,7 @@ public class Main extends Application {
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setTestUi(testUi);
+            fxmlLoader.<MainWindow>getController().setTestUi(duke);
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
