+++ src/main/java/Duke/MainWindow.java
@@ -20,7 +20,7 @@ public class MainWindow extends AnchorPane {
     @FXML
     private TextField userInput;
 
-    private TestUi testUi;
+    private Duke duke;
 
     private Image userImage = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
     private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
@@ -30,8 +30,8 @@ public class MainWindow extends AnchorPane {
         scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
     }
 
-    public void setTestUi(TestUi t) {
-        testUi = t;
+    public void setTestUi(Duke d) {
+        duke = d;
     }
 
     /**
@@ -47,7 +47,7 @@ public class MainWindow extends AnchorPane {
             System.exit(0);
         }
 
-        String response = testUi.getResponse(input);
+        String response = duke.getResponse(input);
         dialogContainer.getChildren().addAll(
                 DialogBox.getUserDialog(input, userImage),
                 DialogBox.getDukeDialog(response, dukeImage)
