--- src/main/java/Duke/Ui.java
@@ -1,12 +1,0 @@
-package duke;
-
-public class Ui {
-
-    /**
-     * Displays greeting message from program.
-     */
-    public void displayGreeting() {
-        System.out.println("Yahallo! I'm Duke\n"
-                + "What can I do for you?\n"); // print greeting message
-    }
-}
