+++ src/test/java/task/TaskTest.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 import org.junit.jupiter.api.Test;
