+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Duke
+Hello! I'm duke
 What can I do for you?
 
 Bye. Hope to see you again soon!
